# publisher: Exclude comment and doctype elements from writeStats