@dirkolbrich 2021-04-20
https://github.com/gohugoio/hugo/pull/8431